chore: seperate checkboxes + list
