+++ src/views/search-results-page/components/filters/components/checkbox.tsx
@@ -1,95 +1,127 @@
 import React from 'react';
 import {
-  Checkbox,
-  Flex,
-  Text,
-  CheckboxProps,
+  Checkbox as ChakraCheckbox,
   Skeleton,
+  Tag,
+  Text,
 } from '@chakra-ui/react';
-import { formatNumber } from 'src/utils/helpers';
+import { FacetTermWithDetails } from 'src/views/search-results-page/components/filters/types';
 
-export interface FiltersCheckboxProps extends CheckboxProps {
-  displayTerm: string; // term used for display
-  value: string; // unique checkbox value.
-  count?: number;
-  isLoading: boolean;
-  property?: string;
-  isCountUpdating?: boolean;
+export interface FilterItem extends FacetTermWithDetails {
+  isHeader?: boolean;
 }
 
-export const FiltersCheckbox: React.FC<FiltersCheckboxProps> = React.memo(
-  ({ displayTerm, count, value, isCountUpdating, isLoading, property }) => {
+// Memoized Checkbox component to prevent unnecessary re-renders
+interface FilterCheckboxProps extends FilterItem {
+  isLoading: boolean;
+  isUpdating?: boolean;
+  colorScheme?: string;
+}
+export const Checkbox: React.FC<FilterCheckboxProps> = React.memo(
+  ({ colorScheme, count, isHeader, isLoading, term, isUpdating, ...props }) => {
+    let label = props.label;
+    let subLabel = '';
+
+    // Display the header label for the group
+    if (isHeader) {
+      return (
+        <Text
+          px={6}
+          fontSize='xs'
+          fontWeight='semibold'
+          lineHeight='shorter'
+          py={1}
+        >
+          {label}
+        </Text>
+      );
+    }
+
+    // Split the label into scientific name and common name if it contains '|'
+    if (term?.includes('|')) {
+      const [scientificName, commonName] = props.label.split(' | ');
+      label = commonName || props.label;
+      subLabel = scientificName;
+    }
+
     return (
-      <Checkbox
+      <ChakraCheckbox
+        value={term}
         w='100%'
-        spacing={3}
-        size='lg'
-        value={value}
-        my={2}
-        sx={{ '.chakra-checkbox__label': { width: '100%' } }}
+        px={6}
+        pr={2}
+        py={1}
+        alignItems='flex-start'
+        _hover={{
+          bg: `${colorScheme}.50`,
+        }}
+        sx={{
+          '>.chakra-checkbox__control': {
+            mt: 1, // to keep checkbox in line with top of text for options with multiple lines
+          },
+          '>.chakra-checkbox__label': {
+            display: 'flex',
+            alignItems: 'center',
+            flex: 1,
+            opacity: count ? 1 : 0.8,
+          },
+        }}
       >
+        {/* Loading skeleton only on load  */}
         <Skeleton
-          width='100%'
-          h={isLoading ? 4 : 'unset'}
-          isLoaded={!isLoading}
+          isLoaded={!isLoading || isUpdating}
+          display='flex'
+          alignItems='center'
           flex={1}
         >
-          <Flex width='100%' alignItems='center'>
-            <Flex
-              w='100%'
-              justifyContent='space-between'
-              alignItems='center'
-              opacity={count ? 1 : 0.7}
-              fontSize='xs'
-              lineHeight={1.5}
-            >
-              {displayTerm && (
-                <Text fontSize='xs' lineHeight={1.5} wordBreak='break-word'>
-                  {(property === 'infectiousAgent' || property === 'species') &&
-                  displayTerm?.includes('|') ? (
-                    <>
-                      {displayTerm
-                        .split(' | ')
-                        .reverse()
-                        .map((term, i) => {
-                          return (
-                            <React.Fragment key={`${term}-${i}`}>
-                              <Text
-                                as='span'
-                                fontWeight={i === 0 ? 'semibold' : 'normal'}
-                              >
-                                {term.charAt(0).toUpperCase() + term.slice(1)}
-                              </Text>
-                              <br />
-                            </React.Fragment>
-                          );
-                        })}
-                    </>
-                  ) : (
-                    <>
-                      {displayTerm.charAt(0).toUpperCase() +
-                        displayTerm.slice(1)}
-                    </>
-                  )}
-                </Text>
-              )}
+          <Text
+            as='span'
+            flex={1}
+            wordBreak='break-word'
+            color='text.heading'
+            fontSize='xs'
+            lineHeight='short'
+            mr={0.5}
+            display='flex'
+            flexDirection='column'
+            fontWeight={subLabel ? 'semibold' : 'normal'}
+          >
+            {label
+              ? label.charAt(0).toUpperCase() + label.slice(1)
+              : 'Loading...'}
+            {subLabel && (
+              <Text
+                as='span'
+                flex={1}
+                wordBreak='break-word'
+                color='text.heading'
+                fontSize='xs'
+                lineHeight='short'
+                fontWeight='normal'
+                mr={0.5}
+              >
+                {subLabel.charAt(0).toUpperCase() + subLabel.slice(1)}
+              </Text>
+            )}
+          </Text>
 
-              {typeof count !== 'undefined' && (
-                <Skeleton
-                  as='span'
-                  px={isCountUpdating ? 3 : 'unset'}
-                  isLoaded={!isCountUpdating}
-                  ml={1}
-                >
-                  <Text as='span' fontWeight='semibold'>
-                    {count ? `${formatNumber(count)}` : '0 '}
-                  </Text>
-                </Skeleton>
-              )}
-            </Flex>
-          </Flex>
+          {/* Display the count of the filter term */}
+          <Tag
+            as='span'
+            className='tag-count'
+            variant='subtle'
+            bg={`${colorScheme}.50`}
+            colorScheme={colorScheme}
+            borderRadius='full'
+            fontSize='xs'
+            alignSelf='flex-start'
+            lineHeight={1.2}
+            size='sm'
+          >
+            {count?.toLocaleString('en-US')}
+          </Tag>
         </Skeleton>
-      </Checkbox>
+      </ChakraCheckbox>
     );
   },
 );
